+++ App/App.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Font } from 'expo';
 
-import Home from './Home';
-import LoadingBackground from './Loading/Background';
+import Screens from './Screens';
+import LoadingBackground from './Screens/Loading/Background';
 
 export default class App extends Component {
   state = {
@@ -21,6 +21,6 @@ export default class App extends Component {
   render() {
     const { fontLoaded } = this.state;
 
-    return this.state.fontLoaded ? <Home /> : <LoadingBackground />;
+    return this.state.fontLoaded ? <Screens /> : <LoadingBackground />;
   }
 }
