+++ App/Screens/Home/Cigarettes/Cigarette/index.js
