+++ App/Screens/Loading/Loading.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 
 import Background from './Background';
-import * as theme from '../utils/theme';
+import * as theme from '../../utils/theme';
 
 export default class Loading extends Component {
   static propTypes = {
