+++ App/Screens/index.js
@@ -1,0 +1,3 @@
+import Screens from './Screens';
+
+export default Screens;
